requirements.txt
boto3
idna
pyyaml
prettytable
